backend/requirements.txt
# Videorama v2.0.0 - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# LLM / OpenAI
openai==1.10.0

# File processing
python-magic==0.4.27
pillow==10.2.0

# Audio/Video processing
yt-dlp==2024.1.7

# JSON
orjson==3.9.12

# Background tasks
celery==5.3.6
redis==5.0.1

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
